pyproject.toml
[project]
name = "pratiche-pareri"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-jazzmin>=3.0.1",
    "djlint>=1.36.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "whitenoise>=6.9.0",
]
[dependency-groups]
# Dipendenze di sviluppo
dev = [
   "django-debug-toolbar>=6.0.0",
]
# Dipendenze aggiuntive per il testing
test = [
    
]
# Dipendenze per lo stage (es. librerie per monitoraggio)
stage = [
    
]
